Select menu implemented.

+++ include/wifi_manager.h
@@ -21,7 +21,7 @@ void wifi_manager_init(void);
 esp_err_t wifi_manager_connect_to_ap(const char* ssid, const char* key);
 void wifi_manager_disconnect(void);
 wifi_manager_status_t wifi_manager_get_status(void);
-esp_err_t wifi_manager_scan(wifi_ap_record_t* ap_info, size_t* count);
+esp_err_t wifi_manager_scan(wifi_ap_record_t* ap_info, uint16_t* count);
 const char* wifi_manager_get_local_ip(void);
 int8_t wifi_manager_get_rssi(void);
